+++ pkg/templates/charts/toggle/cluster-proxy-addon/templates/user-deployment.yaml
@@ -109,6 +109,7 @@ spec:
         ports:
           - name: userport
             containerPort: 9092
+            protocol: TCP
       volumes:
         - name: user-tls-vol
           secret:
