+++ pkg/templates/crds/cluster-proxy-addon/proxy.open-cluster-management.io_managedproxyconfigurations.yaml
@@ -1,6 +1,5 @@
 # Copyright Contributors to the Open Cluster Management project
 
----
 apiVersion: apiextensions.k8s.io/v1
 kind: CustomResourceDefinition
 metadata:
